+++ packages/login-link/resources/views/mail/login-link.blade.php
@@ -6,7 +6,7 @@
 @endphp
 
 <!doctype html>
-<html lang="de">
+<html lang="en">
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">